[hir2mpl]Internal commit msg:         [codecheck const override sign]         [add override]         [ICSL] Clean Code 31         [mapleall] cleancode2  36         [ICSL] Clean Code 28         Fix p70590.c test         [mapleall] cleancode2  24         [WIP] [mapleall] int128: support int128 in FE         [hir2mpl] support neonidx <128 support         [hir2mpl] fix extern -fpie bug     [mapleall]Internal commit msg:         [mapleall] Reconstruct bblayout         [codecheck const override sign]         cg_bugfix_adjust_empty_bb_opt_for_dbg_insn         [mapleall] cleancode2  28         stack protect bug         [mapleall] cleancode2  33         fix sra bug -- need to skip pustatic symbol         [add override]         [ICSL] Clean Code 31         cg_bugfix_temporary_fix_for_array_u64_index         [mapleall] cleancode2  36         [mplcg] do not optimize adrp-label in precfgo         [driver]option specs         [ICSL] Clean Code 28         [ICSL] Clean Code 30         [add override]         [cg]Rollback the patch causes hpf performance degradation         [mapleall] cleancode2  34         [ICSL] Clean Code 29         [live] live add ssa impdef opnd         cg_bugfix_add_adjust_stack_for_addrrr         [mapleall] cleancode2  32         [mplcg] emit visibility for uninitialized global variable & add visibility UT         sra implementation         fix loop vec bug         cg_bugfix_ico_bypass_all_debug_insn_bb         remove bb problem         cg_bugfix_fix_unreachable_bb_in_dead_loop         [mplcg] emit non-default visibility for external functions         refactor hdse DoHDSE interface to get safer code

+++ src/hir2mpl/common/src/hir2mpl_options.cpp
@@ -337,7 +337,7 @@ bool HIR2MPLOptions::ProcessNoMplFile(const maplecl::OptionInterface &) const {
 
 bool HIR2MPLOptions::ProcessDumpLevel(const maplecl::OptionInterface &outputName) const {
   unsigned int arg = outputName.GetCommonValue();
-  FEOptions::GetInstance().SetDumpLevel(arg);
+  FEOptions::GetInstance().SetDumpLevel(static_cast<int>(arg));
   return true;
 }
